# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "مستندات وبسرویس پروژه رستوران",
        "contact": {
            "email": "mhdynasrypak@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth/login_register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "ثبت نام یا ورود",
                "operationId": "89b4fa1e115670ac57505943ec5a5ef9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mobile"
                                ],
                                "properties": {
                                    "mobile": {
                                        "type": "number",
                                        "example": "09055318385"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "داده های معتبر",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "کد به شماره 09055318385 ارسال شد"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "داده‌های نامعتبر",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "سرور با خطا مواجه شد"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/verify_code": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "تایید کد تاییده",
                "operationId": "a44d18767d8bccfc85e286f619fc2fdc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mobile",
                                    "code"
                                ],
                                "properties": {
                                    "mobile": {
                                        "type": "number",
                                        "example": "09055318385"
                                    },
                                    "code": {
                                        "type": "number",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "داده معتبر",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "عملیات ورود با موفقیت انجام شد"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": {
                                                "first_name": null,
                                                "last_name": null,
                                                "mobile": "09055318385",
                                                "status": "active",
                                                "is_admin": "user",
                                                "token": "5|3wNqp8NaONkkpa5oSLEglhhaHbb6NpyEMyfhAF3423c3293e"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "داده‌های نامعتبر",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "سرور با خطا مواجه شد"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}